extracts open ai call to seperate component, adds wind and precip_prob to data
